Upload of FollowAlongFiles zip
Being unsure if the wks went under FollowAlong or are placed in main
PWA folder, I made wks1-4 under main folder / then placed unzipped
goals in wk 1, and put the zip files in the follow along folder.  The
instructions could read either way, this makes the most sense to me, so
went with it.